requirements.txt
# speechbrain
# pydub
numpy
pytube
moviepy
tqdm
pandas
coloredlogs
pillow
opencv-python